pyproject.toml
[tool.poetry]
name = "room-booking"
version = "0.1.0"
description = ""
authors = ["Andrew Medvedev <andmed999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
SQLAlchemy = "^1.4.32"
uvicorn = "^0.17.6"
click = "^8.0.4"
dependency-injector = "^4.38.0"
alembic = "^1.7.6"
psycopg2 = "^2.9.3"
pytest-mock = "^3.7.0"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
ipython = "^8.1.1"
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
autoflake = "^1.4"
pytest-factoryboy = "^2.1.0"
pytest-mock = "^3.7.0"
coverage = "^6.3.2"

[tool.isort]
profile = "black"